On second thought, let's have a standard output.
Related to the design paradigm discussion in the
previous commit, I think it's probably a good
time to highlight why I chose NOT to use JSON or
YAML as the output format: For one thing, the
hurdles and limitations that must be overcome in
memory by making either tags OR files the primary
reference point would not be alleviated by using
a data storage format that is inherently plagued
by the very same issues. (Welcome to many-to-many
data relationships!)

So, instead of cataloging the tags (and which
files they have been applied to) or the files
(and which tags have been applied to them) I
decided to catalog the relationships themselves:
the edges between the nodes, with <==> separating
the tag name from the file name, one line per
association.

Yes, it takes a bit more space. Space-for-speed
(or vise-versa) is always a tradeoff. This format
should be pretty fast (and EASY) to parse though:
want to find all the files with a certain tag?
Use awk or grep. How about all tags applied to a
single file? Awk or grep. It's that simple.